[cleanup] retire deprecated features: incompatible version 2.0.0
* remove sandbox feature
* remove legacy model mgmt
