+++ src/eyepop/index.ts
@@ -55,6 +55,8 @@ export {
     AssetImport,
     ModelExport,
     ModelExportStatus,
+    ModelPrecisionType,
+    ModelFormat,
     QcAiHubExportParams,
     ExportedBy,
     ArtifactType,
@@ -69,13 +71,9 @@ export {
     FileSource,
     MediaStreamSource,
     Source,
-    ModelPrecisionType,
-    ModelFormat,
     PathSource,
     UrlSource,
     AssetUuidSource,
-    ModelInstanceDef,
-    SourcesEntry,
     StreamSource,
     ResultStream,
     VideoMode,
