+++ src/eyepop/worker/worker_options.ts
@@ -9,7 +9,6 @@ export interface WorkerOptions extends Options {
      * Defaults to process.env['EYEPOP_POP_ID'].
      */
     popId?: string | TransientPopId | undefined
-    isSandbox?: boolean
     autoStart?: boolean
     stopJobs?: boolean
     isLocalMode?: boolean
